+++ src/onegov.ballot/onegov/ballot/models/__init__.py
@@ -14,8 +14,10 @@ from onegov.ballot.models.election import ElectionResult
 from onegov.ballot.models.election import List
 from onegov.ballot.models.election import ListConnection
 from onegov.ballot.models.election import ListResult
+from onegov.ballot.models.election import MajorzElection
 from onegov.ballot.models.election import PanachageResult
 from onegov.ballot.models.election import PartyResult
+from onegov.ballot.models.election import ProporzElection
 from onegov.ballot.models.vote import Ballot
 from onegov.ballot.models.vote import BallotResult
 from onegov.ballot.models.vote import ComplexVote
@@ -32,7 +34,9 @@ __all__ = [
     'List',
     'ListConnection',
     'ListResult',
+    'MajorzElection',
     'PanachageResult',
     'PartyResult',
+    'ProporzElection',
     'Vote',
 ]