+++ src/onegov.ballot/onegov/ballot/models/election/majorz_election.py
@@ -1,0 +1,104 @@
+from collections import OrderedDict
+from onegov.ballot.models.election.candidate import Candidate
+from onegov.ballot.models.election.candidate_result import CandidateResult
+from onegov.ballot.models.election.election import Election
+from onegov.ballot.models.election.election_result import ElectionResult
+from sqlalchemy.orm import object_session
+
+
+class MajorzElection(Election):
+    __mapper_args__ = {
+        'polymorphic_identity': 'majorz'
+    }
+
+    def export(self):
+        """ Returns all data connected to this election as list with dicts.
+
+        This is meant as a base for json/csv/excel exports. The result is
+        therefore a flat list of dictionaries with repeating values to avoid
+        the nesting of values. Each record in the resulting list is a single
+        candidate result for each political entity. Party results are not
+        included in the export (since they are not really connected with the
+        lists).
+
+        """
+
+        session = object_session(self)
+
+        ids = session.query(ElectionResult.id)
+        ids = ids.filter(ElectionResult.election_id == self.id)
+
+        results = session.query(
+            CandidateResult.votes,
+            Election.title_translations,
+            Election.date,
+            Election.domain,
+            Election.type,
+            Election.number_of_mandates,
+            Election.absolute_majority,
+            Election.status,
+            Election.counted_entities,
+            Election.total_entities,
+            ElectionResult.group,
+            ElectionResult.entity_id,
+            ElectionResult.elegible_voters,
+            ElectionResult.received_ballots,
+            ElectionResult.blank_ballots,
+            ElectionResult.invalid_ballots,
+            ElectionResult.unaccounted_ballots,
+            ElectionResult.accounted_ballots,
+            ElectionResult.blank_votes,
+            ElectionResult.invalid_votes,
+            ElectionResult.accounted_votes,
+            Candidate.family_name,
+            Candidate.first_name,
+            Candidate.candidate_id,
+            Candidate.elected,
+            Candidate.party,
+        )
+        results = results.outerjoin(CandidateResult.candidate)
+        results = results.outerjoin(CandidateResult.election_result)
+        results = results.outerjoin(Candidate.election)
+        results = results.filter(CandidateResult.election_result_id.in_(ids))
+        results = results.order_by(
+            ElectionResult.group,
+            Candidate.family_name,
+            Candidate.first_name
+        )
+
+        rows = []
+        for result in results:
+            row = OrderedDict()
+            for locale, title in (result[1] or {}).items():
+                row['election_title_{}'.format(locale)] = (title or '').strip()
+            row['election_date'] = result[2].isoformat()
+            row['election_domain'] = result[3]
+            row['election_type'] = result[4]
+            row['election_mandates'] = result[5]
+            row['election_absolute_majority'] = result[6]
+            row['election_status'] = result[7] or 'unknown'
+            row['election_counted_entities'] = result[8]
+            row['election_total_entities'] = result[9]
+
+            row['entity_name'] = result[10]
+            row['entity_id'] = result[11]
+            row['entity_elegible_voters'] = result[12]
+            row['entity_received_ballots'] = result[13]
+            row['entity_blank_ballots'] = result[14]
+            row['entity_invalid_ballots'] = result[15]
+            row['entity_unaccounted_ballots'] = result[16]
+            row['entity_accounted_ballots'] = result[17]
+            row['entity_blank_votes'] = result[18]
+            row['entity_invalid_votes'] = result[19]
+            row['entity_accounted_votes'] = result[20]
+
+            row['candidate_family_name'] = result[21]
+            row['candidate_first_name'] = result[22]
+            row['candidate_id'] = result[23]
+            row['candidate_elected'] = result[24]
+            row['candidate_party'] = result[25]
+            row['candidate_votes'] = result[0]
+
+            rows.append(row)
+
+        return rows