+++ src/onegov.ballot/onegov/ballot/models/vote/__init__.py
@@ -1,7 +1,7 @@
 from onegov.ballot.models.vote.ballot import Ballot
 from onegov.ballot.models.vote.ballot import BallotResult
+from onegov.ballot.models.vote.complex_vote import ComplexVote
 from onegov.ballot.models.vote.vote import Vote
-from onegov.ballot.models.vote.vote import ComplexVote
 
 __all__ = [
     'Ballot',